Refactored Home Page
