+++ src/Pages/Home.jsx
@@ -1,15 +1,14 @@
 import React from "react";
-import { MainLayout } from "../layouts/MainLayout";
-import { HomePageCarousel } from "../components/HomePageCarousel";
-import { ShopByCategory } from "../components/ShopByCategory";
-import { FeaturedCategory } from "../components/FeaturedCategory";
+import { HomePageCarousel } from "../components/HomePage/HomePageCarousel";
+import { ShopByCategory } from "../components/HomePage/ShopByCategory";
+import { FeaturedCategory } from "../components/HomePage/FeaturedCategory";
 import { carousels } from "../FakeData/data";
 import { categories } from "../FakeData/data";
 import { featuredCategories } from "../FakeData/data";
 
 export const Home = () => {
   return (
-    <MainLayout>
+    <>
       <HomePageCarousel carousels={carousels} />
       <ShopByCategory categories={categories} />
       {featuredCategories.map((category) => (
@@ -19,6 +18,6 @@ export const Home = () => {
           items={category.items}
         />
       ))}
-    </MainLayout>
+    </>
   );
 };
