+++ src/Pages/Register.jsx
@@ -1,9 +1,8 @@
 import React from "react";
-import { MainLayout } from "../layouts/MainLayout";
 export const Register = () => {
   return (
-    <MainLayout>
+    <>
       <div>Register</div>
-    </MainLayout>
+    </>
   );
 };
