--- src/Router/App.js
@@ -1,13 +1,0 @@
-import React from "react";
-import { BrowserRouter, Routes, Route, Link, Outlet } from "react-router-dom";
-import { Home } from "../Pages/Home";
-import { Login } from "../Pages/Login";
-import { Register } from "../Pages/Register";
-
-export const App = () => {
-  return (
-    <div>
-      <Home />
-    </div>
-  );
-};
