+++ src/components/ProductCardVertical.css
@@ -1,0 +1,10 @@
+.product-card-vertical {
+  transition: all 140ms linear;
+  padding: 0 0 2rem 0;
+  box-shadow: none;
+}
+
+.product-card-vertical:hover {
+  transform: scale(1.01);
+  box-shadow: 0px 0px 19px 5px rgb(238 243 247);
+}
