+++ src/index.js
@@ -3,11 +3,15 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import "./index.css";
-import { App } from "./Router/App";
+import { App } from "./App";
 import { Login } from "./Pages/Login";
 import { Register } from "./Pages/Register";
 import { Home } from "./Pages/Home";
 
+// The routes are nested in the parent App route, which has the MainLayout Component(holding the navbar and footer)
+// The App component renders the MainLayout and a nested Outlet component for the nested routes to render their elements there.
+// By default at '/' path the index route renders the home page, and at '/login' the login page renders etc.
+// The outlet swaps between the current three routes of home, login and register
 ReactDOM.render(
   <React.StrictMode>
     <BrowserRouter>
@@ -16,6 +20,14 @@ ReactDOM.render(
           <Route index element={<Home />} />
           <Route path="login" element={<Login />} />
           <Route path="register" element={<Register />} />
+          <Route
+            path="*"
+            element={
+              <main style={{ padding: "1rem" }}>
+                <p>There's nothing here!</p>
+              </main>
+            }
+          />
         </Route>
       </Routes>
     </BrowserRouter>
